requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-login==0.6.3
flask-wtf==1.2.1
flask-migrate==4.0.5
flask-limiter==3.5.0
flask-session==0.6.0
werkzeug==3.0.1
wtforms==3.1.1
python-dotenv==1.0.0
gunicorn==21.2.0
flask-socketio==5.3.6
python-socketio==5.11.1
simple-websocket==1.0.0
